+++ src/App.tsx
@@ -1,14 +1,14 @@
 import React from 'react'
 import { Center, Heading, HStack, VStack } from '@chakra-ui/react'
 
-import { UserSettingForm } from './Components'
+import { UserSetting } from './components/UserSetting'
 function App() {
   return (
     <Center p={10}>
       <VStack>
         <Heading>Find ND Filter</Heading>
         <HStack spacing={4}>
-          <UserSettingForm />
+          <UserSetting />
         </HStack>
         <Heading>You need : 100 ND</Heading>
       </VStack>
